+++ CHANGES.md
@@ -14,6 +14,7 @@ Improvements 🙌:
  - Cross-Signing | Setup key backup as part of SSSS bootstrapping (#1201)
  - Cross-Signing | Gossip key backup recovery key (#1200)
  - Show room encryption status as a bubble tile (#1078)
+ - Emoji Verification | It's not the same butterfly! (#1220)
 
 Bugfix 🐛:
  - Missing avatar/displayname after verification request message (#841)